swapped to dialog

+++ installation_scripts/install.sh
@@ -7,7 +7,7 @@ export LINKDOT=${PWD%/*}
 
 sudo pacman -S  go neovim htop firefox xorg-server xorg-xinit xorg-xrdb xorg-xprop \
 		rofi exa pavucontrol tmux pamixer fzf xdg-user-dirs plank sddm \
-		feh git openssh alacritty picom polybar dash xss-lock --needed
+		feh git openssh alacritty picom polybar dash xss-lock dialog --needed
 
 sudo ln -sfT dash /usr/bin/sh
 mkdir -p ~/.config ~/code/aur ~/code/rice
@@ -16,22 +16,31 @@ xdg-user-dir
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 export PATH=$PATH:/home/$USER/.local/bin
 
-read -p "-- install yay as aur helper?" yna
-case $yna in
-           [Yy]* ) git clone https://aur.archlinux.org/yay.git ~/code/aur/yay
-cd ~/code/aur/yay
-makepkg -si
-yay -S ttf-monocraft
-		;;
-	* ) echo "-- you're on your own for the aur, which means fonts won't be autoinstalled.";;
+yay=$(dialog --stdout --inputbox "install yay as aur helper? [y/N]" 0 0) || exit 1
+clear
+shopt -s nocasematch
+if [[ $yay =~ y ]]
+then
+	git clone https://aur.archlinux.org/yay.git ~/code/aur/yay
+        cd ~/code/aur/yay
+        makepkg -si
+        yay -S ttf-monocraft
 
-read -p "-- Would you like to install i3? If you are looking for Qtile say no [y/N] " yna
-case $yna in
-            [Yy]* ) pacman -S i3-wm
-		    ln -sf $LINKDOT/flavours/i3 /home/$USER/.config/
-		                       ;;
-                        * ) echo "-- skipping";;
-esac
+else
+			echo "-- skipping, you are on your own for the aur and fonts."
+fi
+
+i3=$(dialog --stdout --inputbox "install i3? [y/N]" 0 0) || exit 1
+clear
+shopt -s nocasematch
+if [[ $i3 =~ y ]]
+then
+	pacman -S i3-wm
+	ln -sf $LINKDOT/flavours/i3 /home/$USER/.config/
+        
+else
+	echo "-- moving to qtile"
+fi
 
 read -p "-- Would you like to install Qtile?" yna
 case $yna in
